ACCS-1463 - Extend surface should use numbers for edges, not words.

+++ Delcam.ProductInterface.PowerShape/Entities/Surfaces/PSSurface.cs
@@ -841,7 +841,7 @@ namespace Autodesk.ProductInterface.PowerSHAPE
 
             _powerSHAPE.DoCommand("EDIT LIMIT POINT");
             _powerSHAPE.DoCommand("PROFILING " + straightCurved);
-            _powerSHAPE.DoCommand("EDGE " + edge);
+            _powerSHAPE.DoCommand("EDGE " + (int)edge);
             _powerSHAPE.DoCommand("DISTANCE " + length);
             _powerSHAPE.DoCommand("EDIT LIMIT POINT OFF");
         }
